First commit
NOTE: This code is just to display an understanding of good coding style in Php.

The project was based on Admin Panel LTE which is a template generated using the Bootstrap framework.  The Framework is an MVC framework designed by using the following features
1) Autoloading classes using composer - Check RRPhpLibraryV2/Helpers/PhpCrud.php
2) Dependency injection - Check RRPhpLibraryV2/Helpers/PhpCrud.php
3) Namespaces - Check RRPhpLibraryV2/Helpers/PhpCrud.php
4) DAO class - Check models/user.php
5) Single pattern - Check RRPhpLibraryV2/Helpers/Connection.php

The code may raise a lot of questions since it is incomplete.